+++ Altaxo/Base/Graph/Gdi/Shapes/GraphicCollection.cs
@@ -119,7 +119,7 @@ namespace Altaxo.Graph.Gdi.Shapes
 			this.AddRange(g);
 		}
 
-		public void Paint(Graphics g, float Scale, object container)
+		public void Paint(Graphics g, double Scale, object container)
 		{
 			int len = this._items.Count;
 			for (int i = 0; i < len; i++)
